only swap out the promisify signature for TS 4.0+
