+++ types/node/package.json
@@ -11,6 +11,11 @@
             "*": [
                 "ts3.6/*"
             ]
+        },
+        "<=3.9": {
+            "*": [
+                "ts3.9/*"
+            ]
         }
     }
 }
