+++ types/node/ts3.4/index.d.ts
@@ -6,3 +6,4 @@
 /// <reference path="base.d.ts" />
 /// <reference path="assert.d.ts" />
 /// <reference path="globals.global.d.ts" />
+/// <reference path="util.promisify.d.ts" />
