+++ types/node/ts3.9/base.d.ts
@@ -1,0 +1,4 @@
+// NOTE: These definitions support NodeJS and TypeScript 3.5.
+
+// Base definitions for all NodeJS modules that are not specific to any version of TypeScript:
+/// <reference path="../ts3.6/base.d.ts" />
