+++ types/node/ts3.9/node-tests.ts
@@ -1,0 +1,3 @@
+import '../ts3.6/node-tests';
+
+// Nothing special that is not included in the 3.6 tests
