+++ types/node/ts3.9/tslint.json
@@ -1,0 +1,6 @@
+{
+    "extends": "dtslint/dt.json",
+    "rules": {
+        "no-bad-reference": false
+    }
+}
